Create Python version specific Pipfiles
Pipfiles are designed for an application running on a specific Python
minor version in additional to specific versions of dependencies. This
creates the logic to house separate Pipfiles for each supported minor
Python version. The Dockerfile is left to default to the latest version
to make building artifacts locally convenient.